requirements.txt
attrs==19.3.0
aws-cdk.assets==1.27.0
aws-cdk.aws-apigateway==1.27.0
aws-cdk.aws-applicationautoscaling==1.27.0
aws-cdk.aws-autoscaling==1.27.0
aws-cdk.aws-autoscaling-common==1.27.0
aws-cdk.aws-autoscaling-hooktargets==1.27.0
aws-cdk.aws-certificatemanager==1.27.0
aws-cdk.aws-cloudformation==1.27.0
aws-cdk.aws-cloudfront==1.27.0
aws-cdk.aws-cloudwatch==1.27.0
aws-cdk.aws-codebuild==1.27.0
aws-cdk.aws-codecommit==1.27.0
aws-cdk.aws-codepipeline==1.27.0
aws-cdk.aws-ec2==1.27.0
aws-cdk.aws-ecr==1.27.0
aws-cdk.aws-ecr-assets==1.27.0
aws-cdk.aws-ecs==1.27.0
aws-cdk.aws-elasticloadbalancing==1.27.0
aws-cdk.aws-elasticloadbalancingv2==1.27.0
aws-cdk.aws-events==1.27.0
aws-cdk.aws-events-targets==1.27.0
aws-cdk.aws-iam==1.27.0
aws-cdk.aws-kms==1.27.0
aws-cdk.aws-lambda==1.27.0
aws-cdk.aws-logs==1.27.0
aws-cdk.aws-route53==1.27.0
aws-cdk.aws-route53-targets==1.27.0
aws-cdk.aws-s3==1.27.0
aws-cdk.aws-s3-assets==1.27.0
aws-cdk.aws-sam==1.27.0
aws-cdk.aws-secretsmanager==1.27.0
aws-cdk.aws-servicediscovery==1.27.0
aws-cdk.aws-sns==1.27.0
aws-cdk.aws-sns-subscriptions==1.27.0
aws-cdk.aws-sqs==1.27.0
aws-cdk.aws-ssm==1.27.0
aws-cdk.aws-stepfunctions==1.27.0
aws-cdk.core==1.27.0
aws-cdk.cx-api==1.27.0
aws-cdk.region-info==1.27.0
cattrs==1.0.0
# Editable Git install with no remote (covid-figs==0.0.1)
-e /home/shared/janos/covid_figs/covid_figs
importlib-metadata==1.5.0
jsii==0.22.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
publication==0.0.3
py==1.8.1
pyparsing==2.4.6
pytest==5.4.1
python-dateutil==2.8.1
six==1.14.0
typing-extensions==3.7.4.1
wcwidth==0.1.8
zipp==3.1.0
